Diagnostics
If you place this file in the folder in which the GRIME AI executable is installed, and run this batch file from a command prompt, it will records all console output to a file. The file contents can then be used to diagnose issues with GRIME AI, the computer on which it is executing, and/or the web sites it is attempting to communicate with (e.g., USGS, Phenocam, NEON, etc.).

Signed-off-by: John E Stranzl Jr <13703354+JohnStranzl@users.noreply.github.com>