0.0.6.0 (Beta) - Generate a COCO 1.0
This is proof-of-concept (POC) code to take a single mask and apply it to all images or images that each have their own mask and generate a COCO 1.0 annotation file so that a model can be trained. This POC was/is specific to potential functionality needed by NAU (Northern Arizona University) and others.

Signed-off-by: John E Stranzl Jr <13703354+JohnStranzl@users.noreply.github.com>